@@ -1,128 +1,119 @@
 package com.junethewoods.variants.common.block;
 
-import net.minecraft.block.*;
-import net.minecraft.entity.Entity;
-import net.minecraft.item.BlockItemUseContext;
-import net.minecraft.pathfinding.PathType;
-import net.minecraft.state.IntegerProperty;
-import net.minecraft.state.StateContainer;
-import net.minecraft.state.properties.BlockStateProperties;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.tags.FluidTags;
-import net.minecraft.util.Direction;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.shapes.ISelectionContext;
-import net.minecraft.util.math.shapes.VoxelShape;
-import net.minecraft.world.IBlockReader;
-import net.minecraft.world.IWorld;
-import net.minecraft.world.IWorldReader;
-import net.minecraft.world.World;
-import net.minecraft.world.server.ServerWorld;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.level.LevelReader;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.FenceGateBlock;
+import net.minecraft.world.level.block.piston.MovingPistonBlock;
+import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import net.minecraft.world.level.block.state.properties.IntegerProperty;
+import net.minecraft.world.level.pathfinder.PathComputationType;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.VoxelShape;
 import net.minecraftforge.common.FarmlandWaterManager;
-import net.minecraftforge.common.ForgeHooks;
 import net.minecraftforge.common.IPlantable;
 
 import java.util.Random;
 
 public class EndFarmlandBlock extends Block {
-    public static final IntegerProperty MOISTURE = BlockStateProperties.MOISTURE_0_7;
-    protected static final VoxelShape SHAPE = Block.makeCuboidShape(0.0D, 0.0D, 0.0D, 16.0D, 15.0D, 16.0D);
+    public static final IntegerProperty MOISTURE = BlockStateProperties.MOISTURE;
+    protected static final VoxelShape SHAPE = Block.box(0.0D, 0.0D, 0.0D, 16.0D, 15.0D, 16.0D);
 
     public EndFarmlandBlock(Properties properties) {
         super(properties);
-        this.setDefaultState(this.stateContainer.getBaseState().with(MOISTURE, 0));
+        this.registerDefaultState(this.stateDefinition.any().setValue(MOISTURE, 0));
     }
 
-    /**
-     * Update the provided state given the provided neighbor facing and neighbor state, returning a new state.
-     * For example, fences make their connections to the passed in state if possible, and wet concrete powder immediately
-     * returns its solidified counterpart.
-     * Note that this method should ideally consider only the specific face passed in.
-     */
-    public BlockState updatePostPlacement(BlockState stateIn, Direction facing, BlockState facingState, IWorld worldIn, BlockPos currentPos, BlockPos facingPos) {
-        if (facing == Direction.UP && !stateIn.isValidPosition(worldIn, currentPos)) {
-            worldIn.getPendingBlockTicks().scheduleTick(currentPos, this, 1);
+    public BlockState updateShape(BlockState state, Direction direction, BlockState state1, LevelAccessor accessor, BlockPos pos, BlockPos pos1) {
+        if (direction == Direction.UP && !state.canSurvive(accessor, pos)) {
+            accessor.scheduleTick(pos, this, 1);
         }
 
-        return super.updatePostPlacement(stateIn, facing, facingState, worldIn, currentPos, facingPos);
+        return super.updateShape(state, direction, state1, accessor, pos, pos1);
     }
 
-    public boolean isValidPosition(BlockState state, IWorldReader worldIn, BlockPos pos) {
-        BlockState blockstate = worldIn.getBlockState(pos.up());
-        return !blockstate.getMaterial().isSolid() || blockstate.getBlock() instanceof FenceGateBlock || blockstate.getBlock() instanceof MovingPistonBlock;
+    public boolean canSurvive(BlockState state, LevelReader reader, BlockPos pos) {
+        BlockState aboveState = reader.getBlockState(pos.above());
+        return !aboveState.getMaterial().isSolid() || aboveState.getBlock() instanceof FenceGateBlock || aboveState.getBlock() instanceof MovingPistonBlock;
     }
 
-    public BlockState getStateForPlacement(BlockItemUseContext context) {
-        return !this.getDefaultState().isValidPosition(context.getWorld(), context.getPos()) ? Blocks.END_STONE.getDefaultState() : super.getStateForPlacement(context);
+    public BlockState getStateForPlacement(BlockPlaceContext context) {
+        return !this.defaultBlockState().canSurvive(context.getLevel(), context.getClickedPos()) ? Blocks.END_STONE.defaultBlockState() : super.getStateForPlacement(context);
     }
 
-    public boolean isTransparent(BlockState state) {
+    public boolean useShapeForLightOcclusion(BlockState state) {
         return true;
     }
 
-    public VoxelShape getShape(BlockState state, IBlockReader worldIn, BlockPos pos, ISelectionContext context) {
+    public VoxelShape getShape(BlockState state, BlockGetter getter, BlockPos pos, CollisionContext context) {
         return SHAPE;
     }
 
-    public void tick(BlockState state, ServerWorld worldIn, BlockPos pos, Random rand) {
-        if (!state.isValidPosition(worldIn, pos)) {
-            turnToNetherrack(state, worldIn, pos);
+    public void tick(BlockState state, ServerLevel world, BlockPos pos, Random rand) {
+        if (!state.canSurvive(world, pos)) {
+            turnToEndStone(state, world, pos);
         }
     }
 
-    /**
-     * Performs a random tick on a block.
-     */
-    public void randomTick(BlockState state, ServerWorld worldIn, BlockPos pos, Random random) {
-        int i = state.get(MOISTURE);
-        if (!hasWater(worldIn, pos) && !worldIn.isRainingAt(pos.up())) {
-            if (i > 0) {
-                worldIn.setBlockState(pos, state.with(MOISTURE, i - 1), 2);
-            } else if (!hasCrops(worldIn, pos)) {
-                turnToNetherrack(state, worldIn, pos);
+    public void randomTick(BlockState state, ServerLevel world, BlockPos pos, Random rand) {
+        int moisture = state.getValue(MOISTURE);
+        if (!isNearWater(world, pos) && !world.isRainingAt(pos.above())) {
+            if (moisture > 0) {
+                world.setBlock(pos, state.setValue(MOISTURE, moisture - 1), 2);
+            } else if (!isUnderCrops(world, pos)) {
+                turnToEndStone(state, world, pos);
             }
-        } else if (i < 7) {
-            worldIn.setBlockState(pos, state.with(MOISTURE, 7), 2);
+        } else if (moisture < 7) {
+            world.setBlock(pos, state.setValue(MOISTURE, 7), 2);
         }
 
     }
 
-    /**
-     * Block's chance to react to a living entity falling on it.
-     */
-    public void onFallenUpon(World worldIn, BlockPos pos, Entity entityIn, float fallDistance) {
-        if (!worldIn.isRemote && ForgeHooks.onFarmlandTrample(worldIn, pos, Blocks.END_STONE.getDefaultState(), fallDistance, entityIn)) {
-            // Forge: Move logic to Entity#canTrample
-            turnToNetherrack(worldIn.getBlockState(pos), worldIn, pos);
+    public void fallOn(Level world, BlockState state, BlockPos pos, Entity entity, float f) {
+        if (!world.isClientSide && net.minecraftforge.common.ForgeHooks.onFarmlandTrample(world, pos, Blocks.END_STONE.defaultBlockState(), f, entity)) { // Forge: Move logic to Entity#canTrample
+            turnToEndStone(world.getBlockState(pos), world, pos);
         }
 
-        super.onFallenUpon(worldIn, pos, entityIn, fallDistance);
+        super.fallOn(world, state, pos, entity, f);
     }
 
-    public static void turnToNetherrack(BlockState state, World worldIn, BlockPos pos) {
-        worldIn.setBlockState(pos, nudgeEntitiesWithNewState(state, Blocks.END_STONE.getDefaultState(), worldIn, pos));
+    public static void turnToEndStone(BlockState state, Level world, BlockPos pos) {
+        world.setBlockAndUpdate(pos, pushEntitiesUp(state, Blocks.END_STONE.defaultBlockState(), world, pos));
     }
 
-    private boolean hasCrops(IBlockReader worldIn, BlockPos pos) {
-        BlockState plant = worldIn.getBlockState(pos.up());
-        BlockState state = worldIn.getBlockState(pos);
-        return plant.getBlock() instanceof IPlantable && state.canSustainPlant(worldIn, pos, Direction.UP, (IPlantable)plant.getBlock());
+    private boolean isUnderCrops(BlockGetter getter, BlockPos pos) {
+        BlockState plantAbove = getter.getBlockState(pos.above());
+        BlockState state = getter.getBlockState(pos);
+        return plantAbove.getBlock() instanceof IPlantable && state.canSustainPlant(getter, pos, Direction.UP, (IPlantable) plantAbove.getBlock());
     }
 
-    private static boolean hasWater(IWorldReader worldIn, BlockPos pos) {
-        for(BlockPos blockpos : BlockPos.getAllInBoxMutable(pos.add(-4, 0, -4), pos.add(4, 1, 4))) {
-            if (worldIn.getFluidState(blockpos).isTagged(FluidTags.WATER)) {
+    private static boolean isNearWater(LevelReader reader, BlockPos pos) {
+        for(BlockPos blockpos : BlockPos.betweenClosed(pos.offset(-4, 0, -4), pos.offset(4, 1, 4))) {
+            if (reader.getFluidState(blockpos).is(FluidTags.WATER)) {
                 return true;
             }
         }
 
-        return FarmlandWaterManager.hasBlockWaterTicket(worldIn, pos);
+        return FarmlandWaterManager.hasBlockWaterTicket(reader, pos);
     }
 
-    protected void fillStateContainer(StateContainer.Builder<Block, BlockState> builder) {
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
         builder.add(MOISTURE);
     }
 
-    public boolean allowsMovement(BlockState state, IBlockReader worldIn, BlockPos pos, PathType type) {
+    public boolean isPathfindable(BlockState state, BlockGetter getter, BlockPos pos, PathComputationType pathType) {
         return false;
     }
 }
