@@ -1,99 +1,104 @@
 package com.junethewoods.variants.common.item;
 
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.player.ServerPlayerEntity;
-import net.minecraft.item.BowItem;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.ItemUseContext;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.state.Property;
-import net.minecraft.state.StateContainer;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.Hand;
-import net.minecraft.util.Util;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.registry.Registry;
-import net.minecraft.util.text.ChatType;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TranslationTextComponent;
-import net.minecraft.world.IWorld;
-import net.minecraft.world.World;
+import net.minecraft.Util;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Registry;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.chat.ChatType;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.BowItem;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.context.UseOnContext;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.Property;
 
 import javax.annotation.Nullable;
 import java.util.Collection;
 
 public class DebugBowItem extends BowItem {
-    public DebugBowItem(Properties builder) {
-        super(builder);
+    public DebugBowItem(Properties properties) {
+        super(properties);
     }
 
-    public boolean hasEffect(ItemStack stack) {
+    public boolean isFoil(ItemStack stack) {
         return true;
     }
 
-    public boolean canPlayerBreakBlockWhileHolding(BlockState state, World worldIn, BlockPos pos, PlayerEntity player) {
-        if (!worldIn.isRemote) {
-            this.handleClick(player, state, worldIn, pos, false, player.getHeldItem(Hand.MAIN_HAND));
+    public boolean canAttackBlock(BlockState state, Level world, BlockPos pos, Player player) {
+        if (!world.isClientSide) {
+            this.handleInteraction(player, state, world, pos, false, player.getItemInHand(InteractionHand.MAIN_HAND));
         }
         return false;
     }
 
-    public ActionResultType onItemUse(ItemUseContext context) {
-        PlayerEntity playerentity = context.getPlayer();
-        World world = context.getWorld();
-        if (!world.isRemote && playerentity != null) {
-            BlockPos blockpos = context.getPos();
-            this.handleClick(playerentity, world.getBlockState(blockpos), world, blockpos, true, context.getItem());
+    public InteractionResult useOn(UseOnContext context) {
+        Player player = context.getPlayer();
+        Level level = context.getLevel();
+        if (!level.isClientSide && player != null) {
+            BlockPos blockpos = context.getClickedPos();
+            if (!this.handleInteraction(player, level.getBlockState(blockpos), level, blockpos, true, context.getItemInHand())) {
+                return InteractionResult.FAIL;
+            }
         }
-
-        return ActionResultType.func_233537_a_(world.isRemote);
+        return InteractionResult.sidedSuccess(level.isClientSide);
     }
 
-    private void handleClick(PlayerEntity player, BlockState state, IWorld worldIn, BlockPos pos, boolean rightClick, ItemStack stack) {
-        if (player.canUseCommandBlock()) {
+    private boolean handleInteraction(Player player, BlockState state, LevelAccessor reader, BlockPos pos, boolean p_150807_, ItemStack stack) {
+        if (!player.canUseGameMasterBlocks()) {
+            return false;
+        } else {
             Block block = state.getBlock();
-            StateContainer<Block, BlockState> statecontainer = block.getStateContainer();
-            Collection<Property<?>> collection = statecontainer.getProperties();
+            StateDefinition<Block, BlockState> statedefinition = block.getStateDefinition();
+            Collection<Property<?>> collection = statedefinition.getProperties();
             String s = Registry.BLOCK.getKey(block).toString();
             if (collection.isEmpty()) {
-                sendMessage(player, new TranslationTextComponent(this.getTranslationKey() + ".empty", s));
+                message(player, new TranslatableComponent(this.getDescriptionId() + ".empty", s));
+                return false;
             } else {
-                CompoundNBT compoundnbt = stack.getOrCreateChildTag("DebugProperty");
-                String s1 = compoundnbt.getString(s);
-                Property<?> property = statecontainer.getProperty(s1);
-                if (rightClick) {
+                CompoundTag compoundtag = stack.getOrCreateTagElement("DebugProperty");
+                String s1 = compoundtag.getString(s);
+                Property<?> property = statedefinition.getProperty(s1);
+                if (p_150807_) {
                     if (property == null) {
                         property = collection.iterator().next();
                     }
 
-                    BlockState blockstate = cycleProperty(state, property, player.isSecondaryUseActive());
-                    worldIn.setBlockState(pos, blockstate, 18);
-                    sendMessage(player, new TranslationTextComponent(this.getTranslationKey() + ".update", property.getName(), func_195957_a(blockstate, property)));
+                    BlockState blockstate = cycleState(state, property, player.isSecondaryUseActive());
+                    reader.setBlock(pos, blockstate, 18);
+                    message(player, new TranslatableComponent(this.getDescriptionId() + ".update", property.getName(), getNameHelper(blockstate, property)));
                 } else {
-                    property = getAdjacentValue(collection, property, player.isSecondaryUseActive());
+                    property = getRelative(collection, property, player.isSecondaryUseActive());
                     String s2 = property.getName();
-                    compoundnbt.putString(s, s2);
-                    sendMessage(player, new TranslationTextComponent(this.getTranslationKey() + ".select", s2, func_195957_a(state, property)));
+                    compoundtag.putString(s, s2);
+                    message(player, new TranslatableComponent(this.getDescriptionId() + ".select", s2, getNameHelper(state, property)));
                 }
+                return true;
             }
         }
     }
 
-    private static <T extends Comparable<T>> BlockState cycleProperty(BlockState state, Property<T> propertyIn, boolean backwards) {
-        return state.with(propertyIn, getAdjacentValue(propertyIn.getAllowedValues(), state.get(propertyIn), backwards));
+    private static <T extends Comparable<T>> BlockState cycleState(BlockState state, Property<T> comparable, boolean p_40972_) {
+        return state.setValue(comparable, getRelative(comparable.getPossibleValues(), state.getValue(comparable), p_40972_));
     }
 
-    private static <T> T getAdjacentValue(Iterable<T> allowedValues, @Nullable T currentValue, boolean backwards) {
-        return (T)(backwards ? Util.getElementBefore(allowedValues, currentValue) : Util.getElementAfter(allowedValues, currentValue));
+    private static <T> T getRelative(Iterable<T> iterable, @Nullable T p_40975_, boolean p_40976_) {
+        return p_40976_ ? Util.findPreviousInIterable(iterable, p_40975_) : Util.findNextInIterable(iterable, p_40975_);
     }
 
-    private static void sendMessage(PlayerEntity player, ITextComponent text) {
-        ((ServerPlayerEntity)player).func_241151_a_(text, ChatType.GAME_INFO, Util.DUMMY_UUID);
+    private static void message(Player player, Component component) {
+        ((ServerPlayer)player).sendMessage(component, ChatType.GAME_INFO, Util.NIL_UUID);
     }
 
-    private static <T extends Comparable<T>> String func_195957_a(BlockState state, Property<T> propertyIn) {
-        return propertyIn.getName(state.get(propertyIn));
+    private static <T extends Comparable<T>> String getNameHelper(BlockState state, Property<T> comparable) {
+        return comparable.getName(state.getValue(comparable));
     }
 }
