@@ -1,7 +1,7 @@
 package com.junethewoods.variants.common.item;
 
-import net.minecraft.item.Food;
+import net.minecraft.world.food.FoodProperties;
 
 public class F10CompatFoods {
-    public static final Food splash_soph = new Food.Builder().hunger(6).saturation(3).build();
+    public static final FoodProperties SPLASH_SOPH_POTION = new FoodProperties.Builder().nutrition(6).saturationMod(3).build();
 }
