@@ -1,12 +1,12 @@
 package com.junethewoods.variants.common.item;
 
-import net.minecraft.inventory.EquipmentSlotType;
-import net.minecraft.item.ArmorItem;
-import net.minecraft.item.IArmorMaterial;
-import net.minecraft.item.IDyeableArmorItem;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.item.ArmorItem;
+import net.minecraft.world.item.ArmorMaterial;
+import net.minecraft.world.item.DyeableLeatherItem;
 
-public class VariantDyeArmorItem extends ArmorItem implements IDyeableArmorItem {
-    public VariantDyeArmorItem(IArmorMaterial materialIn, EquipmentSlotType slot, Properties builderIn) {
-        super(materialIn, slot, builderIn);
+public class VariantDyeArmorItem extends ArmorItem implements DyeableLeatherItem {
+    public VariantDyeArmorItem(ArmorMaterial material, EquipmentSlot slot, Properties properties) {
+        super(material, slot, properties);
     }
 }
