@@ -2,6 +2,6 @@
 @ParametersAreNonnullByDefault
 package com.junethewoods.variants.common.item;
 
-import mcp.MethodsReturnNonnullByDefault;
+import net.minecraft.MethodsReturnNonnullByDefault;
 
 import javax.annotation.ParametersAreNonnullByDefault;
